requirements/docs.txt
# these are dependencies required to build package documentation
# ought to mirror 'docs' under options.extras_require in setup.cfg
-r install.txt
git+https://github.com/PlasmaPy/plasmapy_sphinx@main#egg=plasmapy_sphinx
docutils >= 0.18.1
ipython
jinja2 >= 3.1.2
nbsphinx >= 0.9.1
numpydoc >= 1.5.0
pillow >= 9.5.0
pygments >= 2.15.0
sphinx >= 6.1.3, <= 7.3.7
sphinx-changelog >= 1.3.0
sphinx-codeautolink >= 0.15.0
sphinx-copybutton >= 0.5.1
sphinx-gallery >= 0.12.2
sphinx-hoverxref >= 1.1.1
sphinx-issues >= 3.0.1
sphinx-notfound-page >= 0.8.3
sphinx-reredirects >= 0.1.1
sphinx_rtd_theme >= 1.2.0
sphinx_tabs >= 3.4.1, <= 3.4.5
sphinxcontrib-bibtex >= 2.5.0
towncrier >= 22.12
tox >= 4.4.0


